# Modern-3D-Site/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlanEventBody {
  /**
   * Free-form description of the event
   * @minLength 3
   */
  user_text: string;
  /** @minimum 0 */
  budget: number;
  /** @minimum 1 */
  guests: number;
  /**
   * Number of days
   * @minimum 1
   */
  duration: number;
  /** @minLength 1 */
  city: string;
}

export interface BudgetCategory {
  name: string;
  amount: number;
  percent: number;
  note?: string;
}

export interface TimelineStep {
  title: string;
  /** Human-friendly timing (e.g. "8 weeks before") */
  when: string;
  description: string;
}

export interface Vendor {
  name: string;
  category: string;
  price: number;
  rating: number;
  city: string;
  contact: string;
  tagline: string;
}

export interface EventPlan {
  eventType: string;
  eventTitle: string;
  summary: string;
  budget: number;
  guests: number;
  duration: number;
  city: string;
  budgetAllocation: BudgetCategory[];
  timeline: TimelineStep[];
  vendors: Vendor[];
  aiAdvice: string[];
}

export interface SavePlanBody {
  plan: EventPlan;
}

export interface SavedPlanSummary {
  id: string;
  eventType: string;
  eventTitle: string;
  city: string;
  budget: number;
  guests: number;
  createdAt: string;
}

export interface SavedPlan {
  id: string;
  createdAt: string;
  plan: EventPlan;
}

export type PlanStatsTopEventTypesItem = {
  eventType: string;
  count: number;
};

export type PlanStatsTopCitiesItem = {
  city: string;
  count: number;
};

export interface PlanStats {
  totalPlans: number;
  totalBudgetPlanned: number;
  totalGuestsPlanned: number;
  topEventTypes: PlanStatsTopEventTypesItem[];
  topCities: PlanStatsTopCitiesItem[];
}