# Modern-3D-Site/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Generate an AI-powered event plan
 */
export const planEventBodyUserTextMin = 3;

export const planEventBodyBudgetMin = 0;

export const PlanEventBody = zod.object({
  user_text: zod
    .string()
    .min(planEventBodyUserTextMin)
    .describe("Free-form description of the event"),
  budget: zod.number().min(planEventBodyBudgetMin),
  guests: zod.number().min(1),
  duration: zod.number().min(1).describe("Number of days"),
  city: zod.string().min(1),
});

export const PlanEventResponse = zod.object({
  eventType: zod.string(),
  eventTitle: zod.string(),
  summary: zod.string(),
  budget: zod.number(),
  guests: zod.number(),
  duration: zod.number(),
  city: zod.string(),
  budgetAllocation: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.number(),
      percent: zod.number(),
      note: zod.string().optional(),
    }),
  ),
  timeline: zod.array(
    zod.object({
      title: zod.string(),
      when: zod
        .string()
        .describe('Human-friendly timing (e.g. \"8 weeks before\")'),
      description: zod.string(),
    }),
  ),
  vendors: zod.array(
    zod.object({
      name: zod.string(),
      category: zod.string(),
      price: zod.number(),
      rating: zod.number(),
      city: zod.string(),
      contact: zod.string(),
      tagline: zod.string(),
    }),
  ),
  aiAdvice: zod.array(zod.string()),
});

/**
 * @summary List recent saved plans
 */
export const ListPlansResponseItem = zod.object({
  id: zod.number(),
  eventType: zod.string(),
  eventTitle: zod.string(),
  city: zod.string(),
  budget: zod.number(),
  guests: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListPlansResponse = zod.array(ListPlansResponseItem);

/**
 * @summary Save a generated plan
 */
export const SavePlanBody = zod.object({
  plan: zod.object({
    eventType: zod.string(),
    eventTitle: zod.string(),
    summary: zod.string(),
    budget: zod.number(),
    guests: zod.number(),
    duration: zod.number(),
    city: zod.string(),
    budgetAllocation: zod.array(
      zod.object({
        name: zod.string(),
        amount: zod.number(),
        percent: zod.number(),
        note: zod.string().optional(),
      }),
    ),
    timeline: zod.array(
      zod.object({
        title: zod.string(),
        when: zod
          .string()
          .describe('Human-friendly timing (e.g. \"8 weeks before\")'),
        description: zod.string(),
      }),
    ),
    vendors: zod.array(
      zod.object({
        name: zod.string(),
        category: zod.string(),
        price: zod.number(),
        rating: zod.number(),
        city: zod.string(),
        contact: zod.string(),
        tagline: zod.string(),
      }),
    ),
    aiAdvice: zod.array(zod.string()),
  }),
});

/**
 * @summary Get a saved plan by id
 */
export const GetPlanParams = zod.object({
  id: zod.string(),
});

export const GetPlanResponse = zod.object({
  id: zod.number(),
  createdAt: zod.coerce.date(),
  plan: zod.object({
    eventType: zod.string(),
    eventTitle: zod.string(),
    summary: zod.string(),
    budget: zod.number(),
    guests: zod.number(),
    duration: zod.number(),
    city: zod.string(),
    budgetAllocation: zod.array(
      zod.object({
        name: zod.string(),
        amount: zod.number(),
        percent: zod.number(),
        note: zod.string().optional(),
      }),
    ),
    timeline: zod.array(
      zod.object({
        title: zod.string(),
        when: zod
          .string()
          .describe('Human-friendly timing (e.g. \"8 weeks before\")'),
        description: zod.string(),
      }),
    ),
    vendors: zod.array(
      zod.object({
        name: zod.string(),
        category: zod.string(),
        price: zod.number(),
        rating: zod.number(),
        city: zod.string(),
        contact: zod.string(),
        tagline: zod.string(),
      }),
    ),
    aiAdvice: zod.array(zod.string()),
  }),
});

/**
 * @summary Aggregate stats across all generated plans
 */
export const GetStatsResponse = zod.object({
  totalPlans: zod.number(),
  totalBudgetPlanned: zod.number(),
  totalGuestsPlanned: zod.number(),
  topEventTypes: zod.array(
    zod.object({
      eventType: zod.string(),
      count: zod.number(),
    }),
  ),
  topCities: zod.array(
    zod.object({
      city: zod.string(),
      count: zod.number(),
    }),
  ),
});